refactor: 클래스 각각 파일로 분리, Defender 클래스 리팩터링(if문 구조 및 출력 형식)
